working on addding representatives to a sidebar which then shows up on the map with the corresponding location. right now the address shows representatives on the side but coordinates not working
